Changes: - resolved the issue with the socialite controller.
